feat(auth): implement RequireMaster component for access control
