+++ apps/console/.env.sample
@@ -1,0 +1,4 @@
+# Backend API URL
+# Development: http://localhost:3001
+# Production: https://your-api-domain.com
+NEXT_PUBLIC_API_URL=http://localhost:3001
