+++ apps/console/app/favicon.ico
