+++ apps/console/eslint.config.mjs
