+++ apps/console/next.config.ts
