+++ apps/console/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "master-console",
+  "name": "console",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -10,7 +10,7 @@
   },
   "dependencies": {
     "@tanstack/react-query": "^5.90.20",
-    "axios": "^1.13.4",
+    "axios": "^1.13.5",
     "next": "16.1.6",
     "react": "19.2.3",
     "react-dom": "19.2.3",
