--- apps/master-console/pnpm-workspace.yaml
@@ -1,3 +1,0 @@
-ignoredBuiltDependencies:
-  - sharp
-  - unrs-resolver
